requirements-phase9.txt
# Phase 9 Testing Dependencies
# Install with: pip install -r requirements-phase9.txt

# Core Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-timeout>=2.1.0
pytest-xdist>=3.0.0  # Parallel test execution

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-pdf==1.0.0  # For basic PDF handling
reportlab>=4.0.0  # For generating test PDFs

# OCR
pytesseract>=0.3.10
easyocr>=1.6.0
paddleocr>=2.7.0  # Fallback OCR
pdf2image>=1.16.0  # Convert PDFs to images for OCR

# NLP/Text Processing
spacy>=3.5.0
transformers>=4.30.0
torch>=2.0.0  # Required by transformers
sentence-transformers>=2.2.0  # For embeddings

# Async & Concurrency
asyncio>=3.4.3
aiofiles>=23.0.0

# Database (for Phase 10)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL adapter

# Testing Utilities
faker>=18.0.0  # Generate fake data
hypothesis>=6.75.0  # Property-based testing
responses>=0.22.0  # Mock HTTP requests

# Performance Testing
pytest-benchmark>=4.0.0

# Code Quality
coverage>=7.0.0
pytest-cov>=4.0.0

# Mocking
unittest-mock>=1.5.0
responses>=0.22.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0  # Progress bars
click>=8.1.0  # CLI utilities
colorama>=0.4.6  # Colored output
